+++ tests/testsuite.sh
@@ -1,5 +1,5 @@
 #! /bin/bash
-TESTDIR=$(dirname $(readlink -f "$0"))
+TESTDIR=$(dirname $(readlink -f "$BASH_SOURCE"))
 MAINDIR=$(dirname "${TESTDIR}")
 
 . "${MAINDIR}/src/templates/suite.tpl"